+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Basedmypy Changelog
 
+## [Unreleased]
+### Fixes
+- `stubtest`: the `__bool__` method of function parameter defaults will no longer be unnecessarily evaluated
+
 ## [2.8.0]
 ### Added
 - ban variance on  `TypeVar` definitions (#817)